requirements.txt
flask==2.0.1
python-dotenv==0.19.0
openai==1.0.0
rich==10.12.0
pydantic>=2.0.0
pytest>=7.0.0 